Create countriesSlice
